Conflicts:
	README.md
	src/protocolsupport/api/ProtocolVersion.java
	src/protocolsupport/protocol/packet/ClientBoundPacket.java
	src/protocolsupport/protocol/serializer/ItemStackSerializer.java
	src/protocolsupport/protocol/typeremapper/block/LegacyBlockData.java
	src/protocolsupport/protocol/typeremapper/chunk/ChunkTransformer.java
	src/protocolsupport/protocol/typeremapper/entity/EntityRemappersRegistry.java
	src/protocolsupport/protocol/typeremapper/itemstack/LegacyItemType.java
	src/protocolsupport/protocol/utils/networkentity/NetworkEntityDataCacheFactory.java
	src/protocolsupport/zplatform/PlatformPacketFactory.java
	src/protocolsupport/zplatform/impl/spigot/SpigotPacketFactory.java

+++ src/protocolsupport/protocol/packet/middle/clientbound/play/MiddleEntityRelMoveLook.java
@@ -12,8 +12,8 @@ public abstract class MiddleEntityRelMoveLook extends MiddleEntity {
 	protected int relX;
 	protected int relY;
 	protected int relZ;
-	protected int yaw;
-	protected int pitch;
+	protected byte yaw;
+	protected byte pitch;
 	protected boolean onGround;
 
 	@Override
